refactor: changed a folder name
